ENH: Use unsigned int vec instead of size_t vec
Switch to using std::vector<unsigned int> instead of std::vector<size_t> for
Index and Size values. This allows them to wrap naturally into python and tcl
lists

+++ Examples/ITKIntegration.cxx
@@ -152,7 +152,7 @@ int main( int argc, char *argv[])
 
   for (int i = 5; i+1 < argc; i+=2)
     {
-    std::vector<size_t> seed;
+    std::vector<unsigned int> seed;
     seed.push_back(atoi(argv[i]));
     seed.push_back(atoi(argv[i+1]));
     segmentationFilter.AddSeed(seed);
